Updated Intro Section
