+++ App.js
@@ -5,8 +5,7 @@ import SmoothNavBar from './src/components/SmoothNavBar';
 import Enums, { NavItem } from './src/utils/Enums';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css';
-import Intro from './src/sections/Intro';
-import Holder from './src/sections/Holder';
+import Intro from './src/sections/intro/Intro';
 import Tech from './src/sections/Tech';
 import PR from './src/sections/pr';
 import Projects from './src/sections/Projects';
