+++ package.json
@@ -17,6 +17,7 @@
     "expo-status-bar": "~1.0.4",
     "i18next": "^20.3.2",
     "i18next-browser-languagedetector": "^6.1.2",
+    "node-sass": "^6.0.1",
     "npm-license-crawler": "^0.2.1",
     "react": "16.13.1",
     "react-bootstrap": "^1.5.2",
