+++ src/components/Http.js
@@ -1,9 +1,11 @@
+import { HttpMethod } from "../utils/Enums"
+
 export default function Http(action, request) {
 
   let url = `https://hanabi-tube.herokuapp.com/${action}`
-  // let localUrl = `http://127.0.0.1:5000/${action}`
+  let localUrl = `http://127.0.0.1:5000/${action}`
 
-  return fetch(url, { request })
+  return fetch(localUrl, { request })
   .then((response) => {
     var res
 
@@ -15,13 +17,22 @@ export default function Http(action, request) {
         "msg": response.statusText,
         'is_error': !response.ok
       }
-      console.log(res)
+      console.error({url: localUrl, res: res})
     }
 
     return new Promise((resolve, reject) => {
       resolve(res)
     })
   })
-  .catch((error) => console.error(error))
+  .catch((error) => console.error({i: localUrl, e: error}))
   
+}
+
+export function MakeRequest(method = HttpMethod.get) {
+  return {
+    method: method,
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  }
 }
